refactor: renamed `query_builder` module to `query`
